#7 fix : nodejs 실행할 app.js 분리, src/public 폴더 구조 변경 (정적 파일은 모두 public에 위치) #25 feat : 사고접수 토글 애니메이션 추가
